+++ app/javascript/controllers/maps_controller.js
@@ -509,6 +509,11 @@ export default class extends BaseController {
         }
       } else if (event.name === 'Tracks') {
         this.handleRouteLayerToggle('tracks');
+      } else if (event.name === 'Areas') {
+        // Show draw control when Areas layer is enabled
+        if (this.drawControl && !this.map.hasControl && !this.map._controlCorners.topleft.querySelector('.leaflet-draw')) {
+          this.map.addControl(this.drawControl);
+        }
       }
 
       // Manage pane visibility when layers are manually toggled
@@ -523,6 +528,11 @@ export default class extends BaseController {
 
         // Manage pane visibility when layers are manually toggled
         this.updatePaneVisibilityAfterLayerChange();
+      } else if (event.name === 'Areas') {
+        // Hide draw control when Areas layer is disabled
+        if (this.drawControl && this.map._controlCorners.topleft.querySelector('.leaflet-draw')) {
+          this.map.removeControl(this.drawControl);
+        }
       }
     });
   }